# openapi_tic_tac_toe/models/game.py
from datetime import date, datetime  # noqa: F401

from typing import List, Dict  # noqa: F401

from openapi_tic_tac_toe.models.base_model import Model
from openapi_tic_tac_toe import util


class Game(Model):
    """NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).

    Do not edit the class manually.
    """

    def __init__(self, game_id=None, game_field=None, game_state=None):  # noqa: E501
        """Game - a model defined in OpenAPI

        :param game_id: The game_id of this Game.  # noqa: E501
        :type game_id: int
        :param game_field: The game_field of this Game.  # noqa: E501
        :type game_field: List[str]
        :param game_state: The game_state of this Game.  # noqa: E501
        :type game_state: str
        """
        self.openapi_types = {
            'game_id': int,
            'game_field': List[str],
            'game_state': str
        }

        self.attribute_map = {
            'game_id': 'game_id',
            'game_field': 'game_field',
            'game_state': 'game_state'
        }

        self._game_id = game_id
        self._game_field = game_field
        self._game_state = game_state

    @classmethod
    def from_dict(cls, dikt) -> 'Game':
        """Returns the dict as a model

        :param dikt: A dict.
        :type: dict
        :return: The Game of this Game.  # noqa: E501
        :rtype: Game
        """
        return util.deserialize_model(dikt, cls)

    @property
    def game_id(self) -> int:
        """Gets the game_id of this Game.

        Game's ID.  # noqa: E501

        :return: The game_id of this Game.
        :rtype: int
        """
        return self._game_id

    @game_id.setter
    def game_id(self, game_id: int):
        """Sets the game_id of this Game.

        Game's ID.  # noqa: E501

        :param game_id: The game_id of this Game.
        :type game_id: int
        """

        self._game_id = game_id

    @property
    def game_field(self) -> List[str]:
        """Gets the game_field of this Game.

        Playing field.  # noqa: E501

        :return: The game_field of this Game.
        :rtype: List[str]
        """
        return self._game_field

    @game_field.setter
    def game_field(self, game_field: List[str]):
        """Sets the game_field of this Game.

        Playing field.  # noqa: E501

        :param game_field: The game_field of this Game.
        :type game_field: List[str]
        """
        allowed_values = ["NONE", "HUMAN", "BOT"]  # noqa: E501
        if not set(game_field).issubset(set(allowed_values)):
            raise ValueError(
                "Invalid values for `game_field` [{0}], must be a subset of [{1}]"  # noqa: E501
                .format(", ".join(map(str, set(game_field) - set(allowed_values))),  # noqa: E501
                        ", ".join(map(str, allowed_values)))
            )

        self._game_field = game_field

    @property
    def game_state(self) -> str:
        """Gets the game_state of this Game.

        Current game's state.  # noqa: E501

        :return: The game_state of this Game.
        :rtype: str
        """
        return self._game_state

    @game_state.setter
    def game_state(self, game_state: str):
        """Sets the game_state of this Game.

        Current game's state.  # noqa: E501

        :param game_state: The game_state of this Game.
        :type game_state: str
        """
        allowed_values = ["CONTINUE", "HUMAN_WIN", "BOT_WIN"]  # noqa: E501
        if game_state not in allowed_values:
            raise ValueError(
                "Invalid value for `game_state` ({0}), must be one of {1}"
                .format(game_state, allowed_values)
            )

        self._game_state = game_state